+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## mesh2scattering
 
+- introduce boundary condition module (m2s.input.bc) to handle non sound hard surfaces ([#53](https://github.com/ahms5/Mesh2scattering/pulls/53))
+
 ### v1.0.1 - (2025-04-25)
 
 - fix copyright in documentation